backend/requirements.txt
# SmartPlayFPL Backend Dependencies

# Web Framework
fastapi==0.115.0

# AI - Claude API
anthropic>=0.39.0
uvicorn[standard]==0.31.0
httpx==0.27.2

# Data Validation
pydantic==2.9.2
pydantic-settings>=2.0.0

# RDF Knowledge Graph (Hybrid_RDF_Scenario)
rdflib>=7.0.0
owlrl>=6.0.2
pyshacl>=0.25.0


# Rate Limiting
slowapi==0.1.9

# Utilities
python-multipart==0.0.12
python-dotenv==1.0.0

# Background Tasks
APScheduler>=3.10.0

# Machine Learning (for expected points model)
scikit-learn>=1.5.0
numpy>=1.26.0
pandas>=2.0.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0  # Async SQLite driver (development)
psycopg2-binary>=2.9.9  # PostgreSQL driver (production - Railway)

# Note: SpaCy is optional - NLP service has regex fallback
# pip install spacy && python -m spacy download en_core_web_sm



